requirements.txt
requests>=2.31
beautifulsoup4>=4.12
icalendar>=5.0
pyyaml>=6.0
geopy>=2.4
python-dateutil>=2.8
playwright>=1.40
lxml>=5.0
tenacity>=8.0
click>=8.0
pytz>=2023.3

# Dev / test
pytest>=7.0
pytest-cov>=4.0
